move types to entities and paths to config
