+++ src/api/products.ts
@@ -1,23 +1,10 @@
 import axios from 'axios';
+import { CategoryType } from 'entities/category';
+import { ProductType } from 'entities/protuct';
 
 const supabaseRoot = 'https://rzknhedkzukvgtstzbxj.supabase.co/rest/v1';
-('/products');
 
-export type CategoryType = {
-  id: number;
-  name: string;
-  image: string;
-};
-
-export type ProductType = {
-  id: number;
-  title: string;
-  price: number;
-  description: string;
-  categories: CategoryType;
-  images: string[];
-};
-
+// NOTE:add pagination to all products request
 // offset: number = 0, limit: number = 9
 const getProducts = async (): Promise<ProductType[]> => {
   const products = await axios.get<ProductType[]>(`${supabaseRoot}/products?select=*,categories(*)'`, {
