+++ src/configs/routerPaths.ts
@@ -1,0 +1,7 @@
+export const routerPaths = {
+  root: '/',
+  products: '/products',
+  productId: ':productId',
+  about: '/about',
+  categories: '/categories',
+};
