+++ src/entities/category.ts
@@ -1,0 +1,5 @@
+export type CategoryType = {
+  id: number;
+  name: string;
+  image: string;
+};
