+++ src/pages/Products/components/Cards/Cards.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from 'react-router-dom';
-import { ProductType } from 'api/products';
 import { Button } from 'components/Button';
 import { Card } from 'components/Card';
+import { ProductType } from 'entities/protuct';
 import styles from './Cards.module.scss';
 
 const Cards = ({ products }: { products: ProductType[] }) => {
